Update DonutNew.cs
error: Newtonsoft.Json.JsonReaderException: "Input string '200.0' is not a valid integer. Path 'amount', line 2, position 17." 
version VKNET: 1.68.0-786

+++ VkNet/Model/GroupUpdate/DonutNew.cs
@@ -21,7 +21,7 @@ namespace VkNet.Model.GroupUpdate
 		/// <summary>
 		/// Цена в рублях
 		/// </summary>
-		public int? Amount { get; set; }
+		public float? Amount { get; set; }
 
 		/// <summary>
 		/// Цена без комиссии (в рублях)
